+++ crates/playwright/src/server/object_factory.rs
@@ -14,9 +14,9 @@
 
 use crate::error::{Error, Result};
 use crate::protocol::{
-    artifact::Artifact, APIRequestContext, Android, Browser, BrowserContext, BrowserType, Dialog,
-    Electron, Frame, LocalUtils, Page, Playwright, Request, ResponseObject, Route, Tracing,
-    WebSocket,
+    APIRequestContext, Android, Browser, BrowserContext, BrowserType, Dialog, Electron, Frame,
+    LocalUtils, Page, Playwright, Request, ResponseObject, Route, Tracing, WebSocket,
+    artifact::Artifact,
 };
 use crate::server::channel_owner::{ChannelOwner, ParentOrConnection};
 use serde_json::Value;
@@ -80,7 +80,7 @@ pub async fn create_object(
                 ParentOrConnection::Parent(_) => {
                     return Err(Error::ProtocolError(
                         "Playwright must have Connection as parent".to_string(),
-                    ))
+                    ));
                 }
             };
 
@@ -100,7 +100,7 @@ pub async fn create_object(
                 ParentOrConnection::Connection(_) => {
                     return Err(Error::ProtocolError(
                         "Browser must have BrowserType as parent".to_string(),
-                    ))
+                    ));
                 }
             };
 
@@ -114,7 +114,7 @@ pub async fn create_object(
                 ParentOrConnection::Connection(_) => {
                     return Err(Error::ProtocolError(
                         "BrowserContext must have Browser as parent".to_string(),
-                    ))
+                    ));
                 }
             };
 
@@ -133,7 +133,7 @@ pub async fn create_object(
                 ParentOrConnection::Connection(_) => {
                     return Err(Error::ProtocolError(
                         "Page must have BrowserContext as parent".to_string(),
-                    ))
+                    ));
                 }
             };
 
@@ -147,7 +147,7 @@ pub async fn create_object(
                 ParentOrConnection::Connection(_) => {
                     return Err(Error::ProtocolError(
                         "Frame must have Page as parent".to_string(),
-                    ))
+                    ));
                 }
             };
 
@@ -161,7 +161,7 @@ pub async fn create_object(
                 ParentOrConnection::Connection(_) => {
                     return Err(Error::ProtocolError(
                         "Request must have Frame as parent".to_string(),
-                    ))
+                    ));
                 }
             };
 
@@ -175,7 +175,7 @@ pub async fn create_object(
                 ParentOrConnection::Connection(_) => {
                     return Err(Error::ProtocolError(
                         "Route must have Frame as parent".to_string(),
-                    ))
+                    ));
                 }
             };
 
@@ -189,7 +189,7 @@ pub async fn create_object(
                 ParentOrConnection::Connection(_) => {
                     return Err(Error::ProtocolError(
                         "Response must have Request as parent".to_string(),
-                    ))
+                    ));
                 }
             };
 
@@ -208,7 +208,7 @@ pub async fn create_object(
                 ParentOrConnection::Connection(_) => {
                     return Err(Error::ProtocolError(
                         "ElementHandle must have Frame as parent".to_string(),
-                    ))
+                    ));
                 }
             };
 
@@ -227,7 +227,7 @@ pub async fn create_object(
                 ParentOrConnection::Connection(_) => {
                     return Err(Error::ProtocolError(
                         "Artifact must have BrowserContext as parent".to_string(),
-                    ))
+                    ));
                 }
             };
 
@@ -241,7 +241,7 @@ pub async fn create_object(
                 ParentOrConnection::Connection(_) => {
                     return Err(Error::ProtocolError(
                         "Dialog must have Page as parent".to_string(),
-                    ))
+                    ));
                 }
             };
 
@@ -285,7 +285,7 @@ pub async fn create_object(
                 ParentOrConnection::Connection(_) => {
                     return Err(Error::ProtocolError(
                         "WebSocket must have Page as parent".to_string(),
-                    ))
+                    ));
                 }
             };